Exit on Program Start, Launch on Program Exit
Ensures that a specific AutoHotkey script (SCR) is running only when a particular program (APP) is not running. Conversely, it ensures that the script is closed whenever the program is active. 
Once the particular program (APP) is closed, the specific AutoHotkey script (SCR) is restarted.
The script achieves this by continuously monitoring the program's process and managing the script's state accordingly

    * Usage Examples: *
01. Restarting a Monitoring Script when a Program is Closed.
02. Auto-starting a Specific Script when a Program is Not Running.
03. Ensure a Script Runs Continuously When a Program is Closed.
04. Switching Between Two Scripts Based on a Running Program.
05. Automatically Launching a Helper Script for a Specific Program.
06. Launching a Backup Script when a Main Application is Not Running.
07. Enabling a Custom Clipboard Manager Only when Certain Programs are Closed.
08. Running a Low-Resource Script when a Heavy Program is Not Active.
09. Auto-Launching a Script when a VPN Client is Not Active.
10. Running a Gaming Optimization Script when the Game is Closed.