Manages script based on specified program's state.
Checks whether a specified program (APP, e.g., smplayer.exe) is running and monitors the presence of another script window (SCR, e.g., Volume Display.ahk). 
Every 200 milliseconds, it checks if the program is running. 
If the program is not found, it will launch the script specified in SCR, unless the script window is already open. 
If the program is found, it checks if the specified script window is open, and if it is, the script window is closed. 
This ensures that the script is only running when the specified program is not active, effectively managing the script window's state based on the program's presence.
    * A bonus script for testing purpose is included *