Triggers a beep and starts the screensaver.
Defines two hotkeys: Ctrl+Enter and Ctrl+NumpadEnter. 
When either of these combinations is pressed, the script emits a brief beep sound (1700 Hz for 75 milliseconds) using the Soundbeep command. 
Following this, the SendMessage command is executed, sending the WM_SYSCOMMAND message (0x0112) with a SC_SCREENSAVE parameter (0xF140) to the Program Manager window. This triggers the screensaver to activate on the system. 
A Return statement ensures no further action is taken after executing the hotkey routine.