Volume indicator with reset & mouse wheel control.
Display and control the system volume in real-time, providing a GUI with a color-coded volume indicator. 
The window remains always on top and updates the volume percentage dynamically, with different colors indicating volume level ranges. Users can reset the volume to the default by clicking the "VOLUME" text in the Gui and adjust the volume via mouse wheel in the taskbar area over the Gui volume display. It also features a close button to terminate the app via a dark red 'X'.
* Option to auto-reload the script every 9 minutes to keep fresh if Un-Commented in the script.